source-code/requirements.txt
pydub
SpeechRecognition
google-cloud-storage
google-cloud-speech
nltk
numpy
networkx
tk
